backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
PyGithub==2.1.1
langchain==0.1.0
langchain-google-genai==0.0.5
aiohttp==3.9.1
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
ruff==0.1.9
